+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## Unreleased
+## 3.0.0-beta.3 - 2022-04-22
 
 ### Changed
-- When editing a Neo field's block types, creating a new block type to edit will show the settings tab, if the field layout tab was previously showing
+- When editing a Neo field's block types, creating a new block type will show the settings tab, if the field layout tab was previously showing
 
 ### Fixed
 - Fixed a bug where creating more than one new block type at once would cause those block types to share a field layout